chore: flushpolicy interval type to uint64
* feat: add shut down public api

* chore: update session handler observer method name

* feat: add remove all method implementation

* feat: add cancel method in async channel implementation

* chore: update lifecycle session wrapper teardown implementation

* chore: update shut down api implementation

* chore: update sample app related to shut down api

* chore: update shut down api test case.

* chore: add missing values in analytics and update shut down test case

* chore: update frequency flush interval type to UInt64

+++ Analytics/Analytics/Policies/FrequencyFlushPolicy.swift
@@ -37,7 +37,7 @@ public final class FrequencyFlushPolicy: FlushPolicy {
     private var flushTimer: Timer?
 
     /// The interval, in milliseconds, at which flush operations occur.
-    private(set) var flushIntervalInMillis: Double
+    private(set) var flushIntervalInMillis: UInt64
 
     /**
      Initializes a new `FrequencyFlushPolicy`.
@@ -45,7 +45,7 @@ public final class FrequencyFlushPolicy: FlushPolicy {
      - Parameter flushIntervalInMillis: The time interval in milliseconds for triggering flushes. Defaults to `Constants.Flush.Interval.default`.
      - Note: The interval is clamped to `Constants.Flush.Interval.min` to prevent excessively short durations.
      */
-    public init(flushIntervalInMillis: Double = Constants.Flush.Interval.default) {
+    public init(flushIntervalInMillis: UInt64 = Constants.Flush.Interval.default) {
         self.flushIntervalInMillis = max(flushIntervalInMillis, Constants.Flush.Interval.min)
     }
